Make field data able to support more than 2B ordinals per segment.
Although segments are limited to 2B documents, there is not limit on the number
of unique values that a segment may store. This commit replaces 'int' with
'long' every time a number is used to represent an ordinal and modifies the
data-structures used to store ordinals so that they can actually support more
than 2B ordinals per segment.

This commit also improves memory usage of the multi-ordinals data-structures
and the transient memory usage which is required to build them (OrdinalsBuilder)
by using Lucene's PackedInts data-structures. In the end, loading the ordinals
mapping from disk may be a little slower, field-data-based features such as
faceting may be slightly slower or faster depending on whether being nicer to
the CPU caches balances the overhead of the additional abstraction or not, and
memory usage should be better in all cases, especially when the size of the
ordinals mapping is not negligible compared to the size of the values (numeric
data for example).

Close #3189

+++ src/main/java/org/elasticsearch/index/fielddata/LongValues.java
@@ -118,7 +118,7 @@ public abstract class LongValues {
             return getValueByOrd(ordinals.getOrd(docId));
         }
 
-        public abstract long getValueByOrd(int ord);
+        public abstract long getValueByOrd(long ord);
 
         @Override
         public final Iter getIter(int docId) {
@@ -127,7 +127,7 @@ public abstract class LongValues {
 
         @Override
         public final long getValueMissing(int docId, long missingValue) {
-            final int ord = ordinals.getOrd(docId);
+            final long ord = ordinals.getOrd(docId);
             if (ord == 0) {
                 return missingValue;
             } else {
@@ -185,7 +185,7 @@ public abstract class LongValues {
         static class Multi implements Iter {
 
             private org.elasticsearch.index.fielddata.ordinals.Ordinals.Docs.Iter ordsIter;
-            private int ord;
+            private long ord;
             private WithOrdinals values;
 
             public Multi(WithOrdinals values) {
